+++ src/layout/App/App.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import logo from './logo.svg';
+import logo from './../../logo.svg';
 import './App.css';
 
 class App extends Component {
